From: Ard Biesheuvel <ard.biesheuvel@linaro.org> Date: Fri, 8 Aug 2014 17:37:36 +0200 Subject: [PATCH 2/4] Add support for cross compilation
This changes the logic that defines ARCH (and HOSTARCH) to take CROSS_COMPILE
into account. Also, $prefix is not assigned, so that the default will be what
is on the path rather than hardcoded in /usr/bin.

This results in the build doing the right thing if CROSS_COMPILE is set in the
environment and no ARCH or prefix options are passed to make, aligning it with
most other CROSS_COMPILE compatible projects.

Change-Id: I18646f1d31955a01f496ecf359d1877ce503dfac
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Nigel Croxon <nigel.croxon@hp.com>