From: Ard Biesheuvel <ard.biesheuvel@linaro.org> Date: Fri, 8 Aug 2014 18:16:59 +0200 Subject: [PATCH 4/4] Add support for 64-bit ARM (AArch64)
This adds support for 64-bit ARM (AArch64) environments. Since there is no
EFI-capable objcopy for this platform, this contains a manually laid out
PE/COFF header using the assembler.

In addition, it includes the relocation bits, some string functions that GCC
assumes are available and other glue to hold it all together.

This can be cross built using

	make CROSS_COMPILE=aarch64-linux-gnu-

Change-Id: Ia60b1eaede8cc5dd4175a325b0e7d7f3bcca8ce9
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Nigel Croxon <nigel.croxon@hp.com>