From: Ard Biesheuvel <ard.biesheuvel@linaro.org> Date: Mon, 11 Aug 2014 15:39:16 +0200 Subject: [PATCH] Add support for 32-bit ARM
This adds support for 32-bit ARM using an approach similar to the one used for
64-bit ARM (AArch64), i.e., it does not rely on an objcopy that is aware of EFI
or PE/COFF, but lays out the entire PE/COFF header using the assembler.

In the 32-bit ARM case (which does not have a division instruction), some code
has been imported from the Linux kernel to perform the division operations in
software.

Change-Id: I84022a1d46592d9992e31afd6a6208f43c9543b1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Nigel Croxon <nigel.croxon@hp.com>